memorial_app/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Login==0.5.0
Flask-Dance==3.0.0
psycopg2-binary==2.9.1
stripe==2.60.0
gunicorn==20.1.0
requests==2.25.1
python-dotenv==0.19.2